Divine Ray:  Fixed damage bugs (hit every tick it was in entity) and added system to multiply damage dealt by number of enemies (OF SAME ENTITY TYPE) it has passed thru.
this means if it passes through 3 pigs, the 3rd pig will take 3 * 8 = 24 damage. If a player were hit by the beam after it hit 3 pigs, the player would take 8 damage. if a second player were hit, they would take 16. if a 4th pig were hit after this, they would take 32 damage.

Fixed bug with activating a second time would not really kill first one, just continued from player location. This would cause the damage to ramp infinitely if passing through enough entities. This is only a problem with extremely short cooldowns, but I hardcoded a stop for this behavior in case we have cooldown shortening abilities in the future.